Making migrations safer for rollback and fixing migrations path in service provider.

+++ migrations/2018_02_27_161729_add_passport_consumer_columns_to_user.php
@@ -14,15 +14,13 @@ class AddPassportConsumerColumnsToUser extends Migration
     public function up()
     {
         Schema::table(config('passport-consumer.user_table'), function (Blueprint $table) {
-
-            if(config('passport-consumer.passport_location') !== 'local'){
+            if (config('passport-consumer.passport_location') !== 'local') {
                 $remoteUserIdentifier = config('passport-consumer.remote_user_identifier');
-                if (!Schema::hasColumn(config('passport-consumer.user_table') ,$remoteUserIdentifier)) {
+                if (!Schema::hasColumn(config('passport-consumer.user_table'), $remoteUserIdentifier)) {
                     $table->string($remoteUserIdentifier)->nullable()->unique();
                 }
                 $table->string('api_token')->nullable();
             }
-
         });
     }
 
@@ -34,21 +32,19 @@ class AddPassportConsumerColumnsToUser extends Migration
     public function down()
     {
         Schema::table(config('passport-consumer.user_table'), function (Blueprint $table) {
-            if(config('passport-consumer.passport_location') !== 'local'){
+            if (config('passport-consumer.passport_location') !== 'local' && Schema::hasColumn(config('passport-consumer.user_table'), 'api_token')) {
                 $table->dropColumn('api_token');
             }
         });
 
         Schema::table(config('passport-consumer.user_table'), function (Blueprint $table) {
 
-            if(config('passport-consumer.passport_location') !== 'local'){
-                if(config('passport-consumer.remove_remote_user_identifier_on_rollback')){
-                    $remoteUserIdentifier = config('passport-consumer.remote_user_identifier');
+            if (config('passport-consumer.passport_location') !== 'local') {
+                $remoteUserIdentifier = config('passport-consumer.remote_user_identifier');
+                if (config('passport-consumer.remove_remote_user_identifier_on_rollback') && Schema::hasColumn(config('passport-consumer.user_table'), $remoteUserIdentifier)) {
                     $table->dropColumn($remoteUserIdentifier);
                 }
             }
-
-
         });
     }
 }
